app/requirements.txt
aiohttp==3.7.2
aiohttp-sse-client==0.1.7
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
crc16==0.1.1
cryptography==3.2.1
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mnemonic==0.19
multidict==5.0.0
pycparser==2.20
PyNaCl==1.3.0
python-dotenv==0.15.0
redis==3.5.3
requests==2.24.0
six==1.15.0
stellar-base-sseclient==0.0.21
stellar-sdk==2.8.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
yarl==1.6.2
